requirements.txt
python-dotenv
PyQt5
openai
google-generativeai
pyaudio
wave
threading
